+++ packages/node-firebird-driver-native/src/test/test.ts
@@ -1,6 +1,6 @@
 import { createNativeClient, getDefaultLibraryFilename } from '../lib';
 
-import { runCommonTests } from 'node-firebird-driver/dist/test/tests';
+import { runCommonTests } from '@knagyorg/node-firebird-driver/dist/test/tests';
 
 const client = createNativeClient(getDefaultLibraryFilename());