+++ packages/node-firebird-driver/src/lib/impl/attachment.ts
@@ -8,6 +8,7 @@ import { AbstractEvents } from './events';
 import {
 	Attachment,
 	Blob,
+	CreateBlobOptions,
 	ExecuteOptions,
 	ExecuteQueryOptions,
 	Events,
@@ -57,6 +58,18 @@ export abstract class AbstractAttachment implements Attachment {
 		await this.preDispose();
 		await this.internalDropDatabase();
 		await this.postDispose();
+	}
+
+	/** Enable/disable cancellation of operations in this attachment. */
+	async enableCancellation(enable: boolean): Promise<void> {
+		this.check();
+		await this.internalEnableCancellation(enable);
+	}
+
+	/** Cancel a running operation in this attachment. */
+	async cancelOperation(forcibleAbort?: boolean): Promise<void> {
+		this.check();
+		await this.internalCancelOperation(forcibleAbort ?? false);
 	}
 
 	/** Executes a statement that uses the SET TRANSACTION command. Returns the new transaction. */
@@ -184,8 +197,8 @@ export abstract class AbstractAttachment implements Attachment {
 		return events;
 	}
 
-	async createBlob(transaction: AbstractTransaction): Promise<AbstractBlobStream> {
-		return await this.internalCreateBlob(transaction);
+	async createBlob(transaction: AbstractTransaction, options?: CreateBlobOptions): Promise<AbstractBlobStream> {
+		return await this.internalCreateBlob(transaction, options);
 	}
 
 	async openBlob(transaction: AbstractTransaction, blob: Blob): Promise<AbstractBlobStream> {
@@ -241,7 +254,9 @@ export abstract class AbstractAttachment implements Attachment {
 
 	protected abstract internalDisconnect(): Promise<void>;
 	protected abstract internalDropDatabase(): Promise<void>;
-	protected abstract internalCreateBlob(transaction: AbstractTransaction): Promise<AbstractBlobStream>;
+	protected abstract internalEnableCancellation(enable: boolean): Promise<void>;
+	protected abstract internalCancelOperation(forcibleAbort: boolean): Promise<void>;
+	protected abstract internalCreateBlob(transaction: AbstractTransaction, options?: CreateBlobOptions): Promise<AbstractBlobStream>;
 	protected abstract internalOpenBlob(transaction: AbstractTransaction, blob: Blob): Promise<AbstractBlobStream>;
 	protected abstract internalPrepare(transaction: AbstractTransaction, sqlStmt: string, options?: PrepareOptions):
 		Promise<AbstractStatement>;