+++ packages/node-firebird-driver/src/lib/index.ts
@@ -90,16 +90,27 @@ export interface FetchOptions {
 	fetchSize?: number;
 }
 
+/** CreateBlobOptions interface. */
+export interface CreateBlobOptions {
+	type?: 'SEGMENTED' | 'STREAM';
+}
+
 /** Attachment interface. */
 export interface Attachment {
 	/** Disconnects this attachment. */
 	disconnect(): Promise<void>;
 
+	/** Enable/disable cancellation of operations in this attachment. */
+	enableCancellation(enable: boolean): Promise<void>;
+
+	/** Cancel a running operation in this attachment. */
+	cancelOperation(forcibleAbort?: boolean): Promise<void>;
+
 	/** Drops the database and release this attachment. */
 	dropDatabase(): Promise<void>;
 
 	/** Creates a blob and return its stream. */
-	createBlob(transaction: Transaction): Promise<BlobStream>;
+	createBlob(transaction: Transaction, options?: CreateBlobOptions): Promise<BlobStream>;
 
 	/** Opens a blob's stream. */
 	openBlob(transaction: Transaction, blob: Blob): Promise<BlobStream>;
@@ -321,6 +332,12 @@ export class Blob {
 	}
 }
 
+export enum BlobSeekWhence {
+	START = 0,
+	CURRENT = 1,
+	END = 2
+}
+
 /** BlobStream class. */
 export abstract class BlobStream {
 	/** Gets the blob's. */
@@ -338,6 +355,9 @@ export abstract class BlobStream {
 
 	/** Cancels the blob's creation. */
 	abstract cancel(): Promise<void>;
+
+	/** Seeks into the blob stream and return the new position. */
+	abstract seek(offset: number, whence?: BlobSeekWhence): Promise<number>;
 
 	/**
 	 * Reads data from the blob and return the number of bytes read or -1 for end-of-stream.